Handle crate names containing hyphens (#31)
* Handle crate names containing '-'

Crates in rust are allowed to contain '-' in their names. But these
dashes are all converted to '_' in the code to not confuse the name with
subtraction. This commit adds this switch into the code, making it no
longer confused for names containing '-'.

* Add test case for hyphen in crate name

This test is the same as the old "simple" test crate, but now the crate
is called "simple-hyphen" instead. This tests that the bundler
handles crate names with hyphens, and expects them to be replaced with
underscores in the code.